share remove all events and render again, modal event background change for not mine events, render function refactoring
