+++ src/js/rest.js
@@ -1,69 +1,12 @@
-import { serverAddress } from "./constants";
-import { calendar } from "../js/calendar/fullCalendar";
 import axios from "axios";
 import $ from "jquery";
+import { serverAddress } from "./constants";
+import { calendar } from "../js/calendar/fullCalendar";
+import { adaptEventsToFullCalendar } from "../js/utils";
 
-var currentUserEvents = [];
-
-function time() {
-  $("header .time").text(calcTime(sessionStorage.zoneDiff));
-}
-setInterval(time, 1000);
-
-//calcTime
-function calcTime(offset) {
-  // create Date object for current location
-  var d = new Date();
-
-  // convert to msec
-  // subtract local time zone offset
-  // get UTC time in msec
-  var utc = d.getTime() + d.getTimezoneOffset() * 60000;
-
-  // create new Date object for different city
-  // using supplied offset
-  var nd = new Date(utc + 3600000 * offset);
-  //ParseDateTime( nd.toLocaleString(), "yyyy-mm-ddThh:mm:ss");
-  var nd2 = new Date(nd);
-  nd2.setHours(nd.getHours());
-  sessionStorage.setItem("currentTime", nd2.toJSON());
-
-  // return time as a string
-  return nd.toLocaleString();
-}
+// calendar.refetchEvents() ------- !
 
 // ------------------------ events ----------------------------------
-// origin
-// const getAllEventsByUser = (userId) => {
-//   const FetchPromise = axios({
-//     method: "GET",
-//     url: serverAddress + "/event/getEventsByUserId",
-//     headers: {
-//       "Content-Type": "application/json",
-//       token: sessionStorage.getItem("token"),
-//     },
-//     data: {},
-//   });
-
-//   FetchPromise.then((res) => {
-//     console.log(res.data.data);
-//     let myNewEvents = res.data.data;
-//     for (let i = 0; i < res.data.data.length; i++) {
-//       myNewEvents[i].start = myNewEvents[i].time;
-//       myNewEvents[i].myDuration = myNewEvents[i].duration;
-//       for (let j = 0; j < myNewEvents[i].roles.length; j++) {
-//         if (myNewEvents[i].roles[j].roleType == "ORGANIZER" && myNewEvents[i].roles[j].user.id == sessionStorage.getItem("userId")) currentUserEvents.push(myNewEvents[i]);
-//       }
-//     }
-//     console.log(currentUserEvents);
-//     //calendar.addEventSource(myNewEvents);///////////
-//   }).catch((error) => {
-//     console.log(error);
-//   });
-// };
-
-// delete event - not implemented
-// cancel event for me  - not implemented
 const getAllEventsByUser = (userId) => {
   const FetchPromise = axios({
     method: "GET",
@@ -74,62 +17,11 @@ const getAllEventsByUser = (userId) => {
     },
     data: {},
   });
-  var myNewEvents;
+  // var myNewEvents;
   FetchPromise.then((res) => {
-    myNewEvents = res.data.data;
-
-    for (let i = 0; i < res.data.data.length; i++) {
-      //research of mostfa and leon DON'T TOUCH OR CHANGE IT
-      //--------------------------------------------------------
-      var updatedZone;
-      var calcZone = sessionStorage.zone.split(":");
-
-      var h = parseInt(calcZone[0] * -1 + parseInt("+04:00"));
-      if (h < 0) {
-        if (h > 10) {
-          updatedZone = "-" + h + ":" + calcZone[1];
-        } else {
-          updatedZone = "-0" + h + ":" + calcZone[1];
-        }
-      } else {
-        if (h > 10) {
-          updatedZone = "+" + h + ":" + calcZone[1];
-        } else {
-          updatedZone = "+0" + h + ":" + calcZone[1];
-        }
-      }
-
-      console.log(updatedZone);
-
-      myNewEvents[i].start = myNewEvents[i].time.split("+")[0] + updatedZone;
-
-      //--------------------------------------------------------
-
-      myNewEvents[i].myDuration = myNewEvents[i].duration;
-      var myHour = new Date(myNewEvents[i].start).getHours();
-      var myMin = new Date(myNewEvents[i].start).getMinutes();
-      var myDuration2 = myNewEvents[i].duration;
-      var calEnd = new Date(myNewEvents[i].start).setHours(myHour + myDuration2, (myHour + myDuration2 - (myHour + parseInt(myDuration2))) * 60 + myMin);
-
-      myNewEvents[i].end = calEnd;
-      myNewEvents[i].resourceId = sessionStorage.getItem("userId");
-
-      if (myNewEvents[i].resourceId == sessionStorage.getItem("userId")) {
-        $(".fc-event").css("background-color", "#D7CDD5").addClass(sessionStorage.getItem("userId"));
-      }
-    }
-    // var resource = [{id: sessionStorage.getItem("userId")}];
-    // myNewEvents.source=resource;
-    console.log("myNewEvents: " + myNewEvents);
-    calendar.addEventSource(myNewEvents); ///////////
-
-    //ADDED class equals to orginaizerid
-    for (let i = 0; i < res.data.data.length; i++) {
-      if (myNewEvents[i].resourceId == sessionStorage.getItem("userId")) {
-        $(".fc-event").css("background-color", "#D7CDD5").addClass(sessionStorage.getItem("userId"));
-        console.log(myNewEvents[i].resourceId);
-      }
-    }
+    const events = adaptEventsToFullCalendar(res.data.data);
+    // console.log(events);
+    calendar.addEventSource(events);
   }).catch((error) => {
     console.log(error);
   });
